Raspberry/requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorzero==2.0
google-api-core==2.16.0
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gpiozero==2.0
httplib2==0.22.0
idna==3.6
oauthlib==3.2.2
paho-mqtt==1.6.1
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
rsa==4.9
uritemplate==4.1.1
urllib3==2.1.0
